Fix converting (all) properties to TFX Properties
I don't know why, but removing the whitespace will sometimes result
in an endless loop always getting a new PsiWhitespace object.

nullable primitive objects need to be SimpleObjectProperties as
the primitive counterpart do not support null.

Lastly, disable "convert all properties to TFX properties" in data classes
as that would result in an empty constructor which is not allowed
for data classes. It would break the data classes functions anyway.

We might want to consider improving the intention action to keep
constructor parameters so the user of that class can create an object
in the same way it was before running this action. Right now we remove the
parameters.
I still think it's wrong for data classes though

fixes #96